feat: integrate GeoFirestore into the home map and add markers.
Still need to find a better way to get the current location of the user.
As my current approach only fetches it once when the composable loads
but it fails the first time when they need to accept permissions and
requires refreshing the composable. Probably need to look into using a
foreground service to refresh the location.

Also if multiple markers are at the same location, only one will be
visible.