feat!: add global progress bar, snackbar for messages and signout user
NOTE: This is a breaking change since I had to move the login and signup
routes under the navigation in Screen.kt as otherwise I am not sure its
possible to navigate from the profile page back to the login page on
signout. This also helps with having a global snackbar for displaying
messages.