perf: fix lag (hopefully) on home page map
The lag was caused by my code that updates the geoquery everytime the
camera position changes, however this was causing the onDocumentChanged
callback to be spammed which was lagging the app as it needs to spawn a
new thread everytime.

The only solution I could come up with was by just not updating the geoquery
and hardcoding it to 100,000km. I may look into updating it if the position
changes by a certain factor but I think this is probably fine for our prototype.